rearrange tests, add file tests
